Refactor: Extract `InitParamField` component
